Factor out ingestion-java-common lib
