+++ ingestion-java-common/bin/mvn
@@ -1,0 +1,20 @@
+#!/bin/bash
+
+# Wrapper script for executing maven via docker, but interacting with
+# the local filesystem. Useful for local development without installing
+# Java and mvn.
+
+set -eo pipefail
+
+cd "$(dirname "$0")/.."
+
+# Create dir to cache maven dependencies if it doesn't already exist.
+mkdir -p ~/.m2
+
+docker run -it --rm \
+    -v ~/.m2:/root/.m2 \
+    -v "$(git rev-parse --show-toplevel)":/root/project \
+    -w /root/project/"$(git rev-parse --show-prefix)" \
+    -e GOOGLE_APPLICATION_CREDENTIALS \
+    maven \
+    mvn "$@"
