+++ ingestion-java-common/checkstyle/suppressions.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0"?>
+
+<!DOCTYPE suppressions PUBLIC
+    "-//Puppy Crawl//DTD Suppressions 1.0//EN"
+    "http://www.puppycrawl.com/dtds/suppressions_1_0.dtd">
+
+<suppressions>
+
+    <!-- google-java-format appears to allow lines over 100 chars in some cases -->
+    <suppress checks="LineLength"
+              files="InputFileFormat.java"/>
+    <suppress checks="LineLength"
+              files="OutputType.java"/>
+
+    <!-- suppress all checks in files we have copied in from other projects -->
+    <suppress checks=".*" files="WriteTables.java" />
+
+    <!-- suppress all checks in the generated directories -->
+    <suppress checks=".*" files=".+[\\\/]generated[\\\/].+\.java" />
+    <suppress checks=".*" files=".+[\\\/]generated-sources[\\\/].+\.java" />
+    <suppress checks=".*" files=".+[\\\/]generated-test-sources[\\\/].+\.java" />
+</suppressions>
