+++ ingestion-java-common/src/main/java/com/mozilla/telemetry/ingestion/common/Env.java
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package com.mozilla.telemetry.ingestion.util;
+package com.mozilla.telemetry.ingestion.common;
 
 import java.util.Optional;
