+++ ingestion-sink/src/test/java/com/mozilla/telemetry/ingestion/SinkIntegrationTest.java
@@ -22,7 +22,7 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.protobuf.ByteString;
 import com.google.pubsub.v1.PubsubMessage;
-import com.mozilla.telemetry.ingestion.util.TestWithPubsubResources;
+import com.mozilla.telemetry.ingestion.common.TestWithPubsubResources;
 import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
