+++ ingestion-sink/src/test/java/com/mozilla/telemetry/ingestion/common/TestWithPubsubResources.java
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package com.mozilla.telemetry.ingestion.util;
+package com.mozilla.telemetry.ingestion.common;
 
 import com.google.api.gax.core.NoCredentialsProvider;
 import com.google.api.gax.grpc.GrpcTransportChannel;
